+++ examples/svelte/src/routes/tests/generic/[topic]/[example]/+page.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+	import { SvelteFlowProvider } from '@xyflow/svelte';
+
+	import Flow from './Flow.svelte';
+
+	export let data: { flowConfig: FlowConfig };
+</script>
+
+<SvelteFlowProvider>
+	<Flow flowConfig={data.flowConfig} />
+</SvelteFlowProvider>