+++ packages/react/src/hooks/useViewportHelper.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import {
   pointToRendererPoint,
-  getTransformForBounds,
+  getViewportForBounds,
   fitView,
   type XYPosition,
   rendererPointToPoint,
@@ -76,50 +76,40 @@ const useViewportHelper = (): ViewportHelperFunctions => {
       },
       fitBounds: (bounds, options) => {
         const { width, height, minZoom, maxZoom, panZoom } = store.getState();
-        const [x, y, zoom] = getTransformForBounds(bounds, width, height, minZoom, maxZoom, options?.padding ?? 0.1);
+        const viewport = getViewportForBounds(bounds, width, height, minZoom, maxZoom, options?.padding ?? 0.1);
 
-        panZoom?.setViewport(
-          {
-            x,
-            y,
-            zoom,
-          },
-          { duration: options?.duration }
-        );
+        panZoom?.setViewport(viewport, { duration: options?.duration });
       },
-      project: (position: XYPosition) => {
-        const { transform, snapToGrid, snapGrid } = store.getState();
-        return pointToRendererPoint(position, transform, snapToGrid, snapGrid);
-      },
-      screenToFlowCoordinate: (position: XYPosition) => {
+      screenToFlowPosition: (position: XYPosition) => {
         const { transform, snapToGrid, snapGrid, domNode } = store.getState();
-        if (domNode) {
-          const { x: domX, y: domY } = domNode.getBoundingClientRect();
 
-          const correctedPosition = {
-            x: position.x - domX,
-            y: position.y - domY,
-          };
-
-          return pointToRendererPoint(correctedPosition, transform, snapToGrid, snapGrid || [1, 1]);
+        if (!domNode) {
+          return position;
         }
 
-        return { x: 0, y: 0 };
+        const { x: domX, y: domY } = domNode.getBoundingClientRect();
+
+        const correctedPosition = {
+          x: position.x - domX,
+          y: position.y - domY,
+        };
+
+        return pointToRendererPoint(correctedPosition, transform, snapToGrid, snapGrid || [1, 1]);
       },
-      flowToScreenCoordinate: (position: XYPosition) => {
+      flowToScreenPosition: (position: XYPosition) => {
         const { transform, domNode } = store.getState();
-        if (domNode) {
-          const { x: domX, y: domY } = domNode.getBoundingClientRect();
 
-          const rendererPosition = rendererPointToPoint(position, transform);
-
-          return {
-            x: rendererPosition.x + domX,
-            y: rendererPosition.y + domY,
-          };
+        if (!domNode) {
+          return position;
         }
 
-        return { x: 0, y: 0 };
+        const { x: domX, y: domY } = domNode.getBoundingClientRect();
+        const rendererPosition = rendererPointToPoint(position, transform);
+
+        return {
+          x: rendererPosition.x + domX,
+          y: rendererPosition.y + domY,
+        };
       },
       viewportInitialized: panZoomInitialized,
     };