+++ packages/svelte/src/lib/index.ts
@@ -87,6 +87,6 @@ export {
   getSmoothStepPath,
   type GetStraightPathParams,
   getStraightPath,
-  getTransformForBounds,
-  getRectOfNodes
+  getViewportForBounds,
+  getNodesBounds
 } from '@xyflow/system';