+++ packages/system/src/xypanzoom/filter.ts
@@ -61,7 +61,7 @@ export function createFilter({
     // if the target element is inside an element with the nopan class, we prevent panning
     if (
       isWrappedWithClass(event, noPanClassName) &&
-      ((!panOnScroll && event.type !== 'wheel') || (panOnScroll && event.type === 'wheel'))
+      (event.type !== 'wheel' || (panOnScroll && event.type === 'wheel' && !zoomActivationKeyPressed))
     ) {
       return false;
     }